requirements.txt
flask
pandas
numpy
scikit-learn
joblib
shap
google-generativeai
opencv-python
pytesseract
SpeechRecognition
networkx